Improve DraftListingDto
